move test summary to reusable code
