+++ ci/github-actions/go-test-summary/summary/tree.go
@@ -1,4 +1,4 @@
-package main
+package summary
 
 import "strings"
